AIMANTRAAS-STU/student-learning-portal/backend/requirements.txt
# AI Mantraas Student Portal - Backend Dependencies
# ===============================================

# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Google Sheets API
google-auth==2.23.4
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.1.1
google-api-python-client==2.111.0

# Data Handling
python-dotenv==1.0.0

# Production Server
gevent==23.9.1
gunicorn==21.2.0

# Static Files (for production)
whitenoise==6.6.0
